Fixes for Windows, GCC7, videos

+++ gemrb/plugins/ACMReader/unpacker.cpp
@@ -167,10 +167,11 @@ int CValueUnpacker::zero_fill(int pass, int /*ind*/)
 	} while (( --i ) != 0);
 	return 1;
 }
+
 int CValueUnpacker::linear_fill(int pass, int ind)
 {
 	int mask = ( 1 << ind ) - 1;
-	short* lb_ptr = buff_middle + ( ( -1ul ) << ( ind - 1 ) );
+	short* lb_ptr = buff_middle + ( ((size_t)-1) << ( ind - 1 ) );
 
 	for (int i = 0; i < subblocks; i++)
 		block_ptr[i * sb_size + pass] = lb_ptr[get_bits( ind ) & mask];
